Add native-dev endpoint for non-generic workers, closes #10
The native-dev endpoint can be used to debug non-generic (i.e., non-dockerized)
workers by allowing the user to specify the service in the querystring. The
user still sends the parameters as JSON inside of a post request, but the
querystring parameter tells it which service to send it to. For example, to
debug a non-generic service called "myservice" the user would send a POST
request to https://api.lanlytics.com/native-dev?service=myservice where the
service parameter is resolved by the backend and tags the message for
myservice.