Change order that messages are pushed to queue
Messages were first pushed to the main queue and a separate entry made for
querying the job status whose key was the job_id. This opens up the possibility
for a job being queried before it has a standalone entry in the database. The
order has now been switched so that standalone entry is entered first then the
job is pushed to the queue. This should avoid cache misses.