Remove VPC creation
We're not going to be able to create a new VPC and still have the
management box (the one this script runs on) be in a security group
that can ssh to everything: there's no way to have security groups
cross VPCs.

New plan: user provisions the VPC, puts management box in it,
runs bursa from there. Bursa figures out what it's running on,
and designates that the management box.